fix: ナビゲーションバーが一度で開かない不具合の修正#339
