--- front/store/navigationDrawer.js
@@ -1,13 +1,0 @@
-export const state = () => ({
-  drawer: false
-})
-
-export const getters = {
-  drawer: (state) => state.drawer
-}
-
-export const mutations = {
-  drawerSet(state) {
-    state.drawer = !state.drawer
-  }
-}
